etl/requirements.txt
python-dotenv
psycopg >= 3.1.0